requirements.txt
Django==6.0
djangorestframework
djangorestframework-simplejwt
# mysqlclient
PyMySQL
cryptography
gunicorn
Pillow
celery
redis